fetchdata added to separate file
